pyproject.toml
[project]
name = "hackathon-anthropic"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["kapso"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
